bmicalcuator.py
Body Mass Index Calculator
This is a simple GUI application written in Python using the Tkinter library that allows the user to calculate their Body Mass Index (BMI). The BMI is a measure of body fat based on height and weight that is used to determine if a person is underweight, normal weight, overweight, obese, or extremely obese.

How to use the application
Open the bmiCalculator file in a Python interpreter or IDE.
Run the program by clicking the green "play" button or pressing F5.
Enter your weight and height in kilograms and centimeters, respectively.
The BMI will be calculated and displayed in the window.
The BMI will be color-coded to indicate whether you are underweight, normal weight, overweight, obese, or extremely obese.
The BMI classification will also be displayed.
Requirements
In order to run this application, you will need:

Python 3.6 or later
The Tkinter library.